requirements.txt
flask==3.0.0
langchain-community
langchain-text-splitters==0.0.1
langchain-google-genai==0.0.6
langchain
chromadb==0.4.18
python-dotenv==1.0.0